*Fixed RightClickMenu being a bit too far to the left on the lower part *Fixed bug that prevented copying/moving of empty directory (thx dude) *Added a netconfig information that can get all network information of the wii (isn't used right now) *Added a few functions for text converting *Changed the screenwidth by a bit *Lots of changes (mostly fixes) to the TextEditor, created a Text class (might be new bugs in it, report if you find some) *TextEditor is now working completely on UTF8 *Made OnScreenKeyboard complete UTF8 compatible. *Added possibility to change the Keys on the OnScreenKeyboard through the LanguageFile *FreeTypeGX modification. Huge speedup in it now especially for long text. *Added wstring class by Hibernatus to the source
NOTE to Keyboardtranslation:
The translation might include following characters "qwertyuiop[]\\". These are the charecters in the order they appear on the Keyboard. Some characters require in C/C++ to have a precharecter like '\' which is than '\\' as you see above. The gettext lib seems to write both of these to the language file but is not reading both of them. That means you have to change it to one '\'. Same goes for '%%' to '%'. Otherwise the app won't find the translation.
For example:
"qwertyuiop[]\\"
becomes:
"qwertyuiop[]\"

This breaks though the ability to compile the source than. I'll try to find a solutition to it. In the meantime you can use those language files locally.