Experemental Commit: There were a lot of changes in this rev. I can't remember all but mainly:
*Fix for THP crash on unsupported/unloadable files
*Few fixes in the TextEditor
*Main AudioCore change. A lot of changes in the audio playback.
*Playback of audio now directly from the device possible. This makes the playback of any filesize possible for the supported formats. There is a settings to load the files to memory first and play them from there (old way).
*Tooltip Class added (no tooltips added to buttons yet, a bit lazy :P)
*Fix for googlecode update


NOTE: The audio changes were a bit instable for me. That is why this is an experemental commit. 
Please report if the playback is ok or not for you and also what else is better and what is worse. In the Settings there are also a few soundbuffer options which you could play with to adjust and report the best configs for you. If these new modifications to the core of sound playback won't work out, they might get reverted.

NOTE2: You will need to update to this revision manually. Due to a few changes from google on the googlecode site the autoupdater doesn't work anymore. This is fixed in the new revision for future releases.