PDF Changes:    *Improved PDF error handling. No need to go out of app now on a fatal error.    *More memory now available on PDF processing.    *Changes in the libmupdf are uploaded to the branches
*Fixed bug in ProgressBar when filesize is bigger than 2GB

Initial Commit of new "Minimize ProgressBar" Feature. You can minimize the process and let it run in the background. Please report new bugs if you find something.

Note 1: If you have a copy process running and copy/paste something else you can append this to the transfer queue. It will automaticly follow after the first transfer is finished.

Note 2: The Thread Priority is configurable in the settings for both modes "Normal" and "Background" Mode. You can configure the copy speed with it.

Note 3: The Process is minimized to the Taskbar. I didn't have the proper images so for now there is only a clickable text like "Copying..." or "Moving..." which brings you back the ProgressBar. As soon as i got the images i will do a proper taskbar button.