*A few code optimization...nothing really new *Compiled with new SVN libogc (includes a few USB optimizations for slow starting USB devices) *Language files updated
Note: I am compiling this Rev with libogc SVN R4306 and R4153 of libogc which was having better compatibility for USB1.0 drives/sticks/hubs. Try using the normal revision of WiiXplorer and if your device is not detected try the revision compiled with libogc R4153. You will have to download the second one manually.